refactor(list): Barrel file renaming
