+++ projects/igniteui-angular/src/lib/grids/filtering/excel-style/excel-style-search.component.ts
@@ -14,7 +14,7 @@ import { IgxInputDirective } from '../../../directives/input/input.directive';
 import { DisplayDensity } from '../../../core/density';
 import { IgxForOfDirective } from '../../../directives/for-of/for_of.directive';
 import { FilterListItem } from './grid.excel-style-filtering.component';
-import { IgxListComponent } from '../../../list';
+import { IgxListComponent } from '../../../list/public_api';
 
 @Directive({
     selector: '[igxExcelStyleLoading]'
